refactor(renderer): extract the HDR scene pass into scene_pass.rs
Cluster 8 (the big one): sky-view LUT refresh, sky + immediate-mode 3D
batch + scene-graph render into the HDR MRT set, planar reflections,
and the opaque material pass on the inner render graph all move into
record_hdr_scene_pass (266 lines). has_3d rebinds inside the method
(same predicate as the upload step just before the call; vertices_3d
untouched between); encoder reborrows feed the inner FrameCtx. Goldens
pixel-identical. mod.rs ratchets 12359 -> 12100.
